Adopt localstorage, refactor into lib/restore-last-path
