+++ client/controller/index.node.js
@@ -43,7 +43,3 @@ export function redirectLoggedIn( { isLoggedIn, res }, next ) {
 
 	next();
 }
-
-export function restoreLastSession() {
-	return false;
-}
